Fixes drt 2014
* multiscaleimage.cpp: Fix the type of the left shit operations to
avoid overflows. Make it an inline function for easier maintenance.
* tilesource.[h|cpp]: Fix the type of the image width and height,
should be bigger than int.
* System.Windows.Media/MultiScaleTileSource.cs: Fix cast.

* Mono/GeneratedPInvokes.cs, cbinding.[h|cpp]: regenerate

svn path=/trunk/moon/; revision=139588

+++ class/System.Windows/System.Windows.Media/MultiScaleTileSource.cs
@@ -20,11 +20,11 @@ namespace System.Windows.Media
 	public abstract partial class MultiScaleTileSource : DependencyObject
 	{		
 		internal long ImageWidth {
-			get { return NativeMethods.multi_scale_tile_source_get_image_width (this.native); }
+			get { return (long)NativeMethods.multi_scale_tile_source_get_image_width (this.native); }
 			set { NativeMethods.multi_scale_tile_source_set_image_width (this.native, value); }
 		}
 		internal long ImageHeight {
-			get { return NativeMethods.multi_scale_tile_source_get_image_height (this.native); }
+			get { return (long)NativeMethods.multi_scale_tile_source_get_image_height (this.native); }
 			set { NativeMethods.multi_scale_tile_source_set_image_height (this.native, value); }
 		}
 		internal int TileWidth {
